move SourceCode class to utils.py
