+++ elpy/jedibackend.py
@@ -11,10 +11,6 @@ import sys
 import traceback
 import re
 
-from array import array
-from io import StringIO
-from bisect import bisect_right
-
 from typing import List, Optional, Union, NamedTuple, Any, NoReturn
 from pathlib import Path
 import jedi
@@ -22,6 +18,8 @@ import jedi
 
 from elpy import rpc
 from elpy.rpc import Fault, Result
+from elpy.utils import SourceCode
+
 
 class NameResult(Result):
     name: str
@@ -63,7 +61,6 @@ class RefactoringResult(Result):
 
 # in case pkg_resources is not properly installed
 # (see https://github.com/jorgenschaefer/elpy/issues/1674).
-
 try:
     from pkg_resources import parse_version
 except ImportError:  # pragma: no cover
@@ -71,6 +68,7 @@ except ImportError:  # pragma: no cover
         raise Fault("`pkg_resources` could not be imported, "
                     "please reinstall Elpy RPC virtualenv with"
                     " `M-x elpy-rpc-reinstall-virtualenv`", code=400)
+
 
 class JediBackend(object):
     """The Jedi backend class.
@@ -83,7 +81,6 @@ class JediBackend(object):
     name = "jedi"
 
     def __init__(self, project_root, environment_binaries_path):
-            
         self.project_root = project_root
         self.environment = None
         if environment_binaries_path is not None:
@@ -94,7 +91,7 @@ class JediBackend(object):
                 Fault(message=str(e))
         self.completions = {}
         sys.path.append(project_root)
-    
+
     def rpc_get_completions(self, filename, source, offset):
         src = SourceCode(filename, source)
         line, column = src.get_pos(offset)
@@ -177,10 +174,8 @@ class JediBackend(object):
             return None
         return (str(loc.module_path), offset)
 
-
     def rpc_get_assignment(self, filename, source, offset):
         raise Fault("Obsolete since jedi 17.0. Please use 'get_definition'.")
-
 
     def rpc_get_calltip(self, filename, source, offset):
         line, column = pos_to_linecol(source, offset)
@@ -203,6 +198,7 @@ class JediBackend(object):
         Return the current function calltip or its oneline docstring.
 
         Meant to be used with eldoc.
+
         """
         # Try to get a oneline docstring then
         docs = self.rpc_get_oneline_docstring(filename=filename,
@@ -330,7 +326,7 @@ class JediBackend(object):
                 offset = other_src.get_offset(name.line, name.column)
             result.append(NameResult.from_name(name, offset))
         return result
-        
+
     def rpc_get_rename_diff(self, filename, source, offset, new_name):
         """Get the diff resulting from renaming the thing at point"""
         src = SourceCode(filename, source)
@@ -391,91 +387,6 @@ class JediBackend(object):
             return RefactoringResult.fail()
         else:
             return RefactoringResult.from_refactoring(ref)
-
-
-# From the Jedi documentation:
-#
-#   line is the current line you want to perform actions on (starting
-#   with line #1 as the first line). column represents the current
-#   column/indent of the cursor (starting with zero). source_path
-#   should be the path of your file in the file system.
-class Pos(NamedTuple):
-    line: int
-    col: int
-
-class SourceCode:
-    _source: Optional[str]
-    _path: Optional[Path]
-    _line_offsets: Optional[array[int]]
-    
-    def __init__(
-            self, path: Union[None, str, Path], source: Optional[str] = None,
-            line_start: int = 1, col_start: int = 0):
-        if path is not None:
-            self._path = Path(path)
-        self._source = source
-        self._line_offsets = None
-        self._line_start = line_start
-        self._col_start = col_start
-        
-    def get_source(self):
-        if self._source is None:
-            with open(self._path, 'r') as fh:
-                self._source = fh.read()
-        return self._source
-
-    def __str__(self):
-        return self.get_source()
-
-    @property
-    def path(self) -> Optional[Path]:
-        return self._path
-
-    def get_pos(self, offset: int) -> Pos:
-        """Return a tuple of line and column for offset pos in text.
-
-        Lines are one-based, columns zero-based.
-
-        This is how Jedi wants it. Don't ask me why.
-        
-        """
-        idx = self._get_line_offsets()
-        line_num_from_zero = bisect_right(idx, offset) - self._line_start
-        line_offset = idx[line_num_from_zero]
-        return Pos(line_num_from_zero + self._line_start,
-                   offset - line_offset + self._col_start)
-
-    def get_offset(self, line: int, col: int):
-        """Return the offset of this line and column in text.
-        
-        Lines are one-based, columns zero-based.
-        
-        This is how Jedi wants it. Don't ask me why.
-        
-        """
-        line_offsets = self._get_line_offsets()
-        line_from_zero = line - self._line_start
-        col_from_zero = col - self._col_start
-        try:
-            cur_line_offset = line_offsets[line_from_zero]
-        except IndexError:
-            raise ValueError(f"Text does not have {line} lines.")
-        offset =  cur_line_offset + col_from_zero
-        next_line_offset = line_offsets[line_from_zero + 1]
-        if not offset < next_line_offset:
-            raise ValueError(
-                f"Line {line} column {col} is not within the text")
-        return offset
-    
-    def _get_line_offsets(self) -> array[int]:
-        if self._line_offsets is None:
-            self._line_offsets = array('I')
-            self._line_offsets.append(0)
-            curr_line_offset = 0
-            for line in StringIO(self.get_source()):
-                curr_line_offset += len(line)
-                self._line_offsets.append(curr_line_offset)
-        return self._line_offsets
 
 
 def pos_to_linecol(text, pos):
